♻️ Refactor: client & server 컴포넌트 명시
