+++ src/common/components/Cta/TextCta/TextCta.client.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import PlusIcons from '@icons/PlusIcons/PlusIcons';
 import { PlusIconsState } from '@icons/PlusIcons/PlusIcons.types';
